feat: add metadata to Job events (#63)
Add Job metadata to APM via custom context:
* job_id: the job identifier.
* attempts: the number of times the job has been attempted.
* max_tries: the number of times to attempt a job.
* connection_name: the name of the connection the job belongs to.
* queue_name: name of the queue the job belongs to.

Important: Custom metadata adds non-indexed, custom contextual information to transactions and errors. Non-indexed means the data is not searchable or aggregatable in Elasticsearch, and you cannot build dashboards on top of the data. This also means you don’t have to worry about mapping explosions, as these fields are not added to the mapping.

https://www.elastic.co/guide/en/apm/get-started/current/metadata.html#custom-fields