Remove injection of Agent class into EventDataCollector classes
The FrameworkCollector is created during the Laravel 'register' phase, which was forcing the Agent to be constructed as well. This occurred before other service providers had completed registering and therefore prevented modifying the Agent construction through the AgentBuilder abstraction.

Collectors are now tagged during 'register' and given to the Agent, rather than being created within the Agent.

The Agent is given to the Collector objects during the add process to meet the need of the JobCollector interacting with it.