Chore/migrate php agent (#66)
* Update to nipwaayoni/elastic-apm-php-agent 7.3 release

* Refactor Agent

The parent class now has a final constructor and supports initialization through an explicit overridden method.

* Refactor ServiceProvider

The updated Agent package includes an AgentBuilder to facilitate constructing the agent. Use it to fulfill the Agent implementation in the service container.

Some changes were made to the Agent configuration settings and environment values as well.

The Agent no longer relies on the class __destruct() method to call send(). We need to explicitly call send() on the agent which is best done in an App::terminating() registered callback. This is executed after the framework has sent the response to the consumer.