chore: separate collector init from agent creation (#70)
* Update to nipwaayoni/elastic-apm-php-agent 7.3 release

* Refactor Agent

The parent class now has a final constructor and supports initialization through an explicit overridden method.

* Refactor ServiceProvider

The updated Agent package includes an AgentBuilder to facilitate constructing the agent. Use it to fulfill the Agent implementation in the service container.

Some changes were made to the Agent configuration settings and environment values as well.

The Agent no longer relies on the class __destruct() method to call send(). We need to explicitly call send() on the agent which is best done in an App::terminating() registered callback. This is executed after the framework has sent the response to the consumer.

* Remove injection of Agent class into EventDataCollector classes

The FrameworkCollector is created during the Laravel 'register' phase, which was forcing the Agent to be constructed as well. This occurred before other service providers had completed registering and therefore prevented modifying the Agent construction through the AgentBuilder abstraction.

Collectors are now tagged during 'register' and given to the Agent, rather than being created within the Agent.

The Agent is given to the Collector objects during the add process to meet the need of the JobCollector interacting with it.

* Extend AgentBuilder

Extending the AgentBuilder allows for capturing custom Agent creation behaviors and exposing them in a consistent way.

* Remove Agent::setRequestStartTime

This was added in an earlier version and the start time is now expected in the constructor.

* Fix code style issues

* Encapsulate decision to include HTTP collector

The methods have also been renamed to more accurately reflect their purpose.

* Reset composer dependencies

* Update only nipwaayoni/elastic-apm-php-agent