chore: support ELASTIC_APM environment variables (#78)
* Redo package config to support ELASTIC_APM environment variables

Support for environment variables is planned for the next release of the Agent package. Config object arguments will always override environment variables, so we should only pass non-null keys.

Switch to base Span and remove LazySpan so we don’t have to maintain that class.

Use real Transaction in place of mock when testing RecordTransaction middleware to avoid calling protected methods of the partially mocked Transaction by casting the Transaction to json and back to an array.

* Add ApmAgent facade

* Replace use of 'config()' helper with injected config repository object

Unit testing the Agent was difficult with the use of Laravel's 'config()' helper. Replacing that with an injected config repository, similar to how the event collectors get the config, solves that problem. This required changes back through the AgentBuilder and ServiceProvider.

* Inject Agent into Services and add tests for ApmAgentService

* Fix unit test for Agent constructor change